delete refresh button vehicle - add automatically refresh

+++ CarRentalApp/ManageVehicleListing.cs
@@ -21,31 +21,6 @@ namespace CarRentalApp
 
         private void ManageVehicleListing_Load(object sender, EventArgs e)
         {
-            //Select * From Cars
-            //var cars = _db.Cars.ToList(); 
-
-            //Select Id as CarId, name as CarName from Cars
-            //var cars = _db.Cars
-            //    .Select(q => new { CarId = q.Id, CarName = q.Make })
-            //    .ToList();
-            //gvVehicleList.DataSource = cars;
-            //gvVehicleList.Columns[0].HeaderText = "ID";
-            //gvVehicleList.Columns[1].HeaderText = "Make";
-
-            //var cars = _db.Cars
-            //    .Select(q => new
-            //    {
-            //        Make = q.Make,
-            //        Model = q.Model,
-            //        Vin = q.Vin,
-            //        Year = q.Year,
-            //        LicensePlateNumber = q.LicensePlateNumber,
-            //        q.Id
-            //    })
-            //    .ToList();
-            //gvVehicleList.DataSource = cars;
-            //gvVehicleList.Columns[4].HeaderText = "License Plate Number";
-            //gvVehicleList.Columns[5].Visible = false;
             try
             {
                 PopulateGrid();
@@ -58,7 +33,7 @@ namespace CarRentalApp
 
         private void btnAddCar_Click(object sender, EventArgs e)
         {
-            AddEditVehicle addEditVehicle = new AddEditVehicle();
+            AddEditVehicle addEditVehicle = new AddEditVehicle(this);
             addEditVehicle.MdiParent = this.MdiParent;
             addEditVehicle.Show();
         }
@@ -77,7 +52,7 @@ namespace CarRentalApp
                 var car = _db.Cars.FirstOrDefault(q => q.Id == id);
 
                 //launch AddEditVehicle window with data
-                var addEditVehicle = new AddEditVehicle(car);
+                var addEditVehicle = new AddEditVehicle(car, this);
                 addEditVehicle.MdiParent = this.MdiParent;
                 addEditVehicle.Show();
             }
@@ -97,11 +72,15 @@ namespace CarRentalApp
                 //query database for record
                 var car = _db.Cars.FirstOrDefault(q => q.Id == id);
 
-                //delete vehicle from table
-                _db.Cars.Remove(car);
-                _db.SaveChanges();
-
-                gvVehicleList.Refresh();  
+                DialogResult dr = MessageBox.Show("Are you sure you want to delete this record?",
+                    "Delete", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Warning);
+                if (dr == DialogResult.Yes)
+                {
+                    //delete vehicle from table
+                    _db.Cars.Remove(car);
+                    _db.SaveChanges();
+                }
+                PopulateGrid();
             }
             catch (Exception ex)
             {
@@ -115,9 +94,11 @@ namespace CarRentalApp
         {
             //Simple Refresh Option
             PopulateGrid();
+            gvVehicleList.Update();
+            gvVehicleList.Refresh();
         }
 
-        private void PopulateGrid()
+        public void PopulateGrid()
         {
             var cars = _db.Cars
                 .Select(q => new
